led cycle not working
we made an led blink function that calculates the led state in a fancy way.
for some reason it is not blinking the led on the board or the external led.
Depending on how we print stuff, the modulo 500 is actually modulo'ing 5000.
Regardless, the led does not really appear to be changing (it may be blinking very fast?)
Next test is have the led blink slowly.