Got main state machine working
We now have 4 states in our loop.
First we go towards an object,
then we turn away until we no longer see the object,
then we turn until we see the object again
then we change our target to the next object and start searching for it.
Once found, we go back to going towards step