Fix pixiblock indexing error
We encountered a problem where we detected unreasonably large areas when using a target = 3
The issue had to do with that the targets were 1 indexed and we were using this
index, rather than subtracting one, causing an overflow memory error.
We still see the unrelated 'cs error's for some reason.